moved our unofficial test into an official test
