+++ src/Examples.jl
@@ -1,7 +1,7 @@
-# sorts M, A
-# operations m: M,M -> M
-#            a: M,A -> A
-# equations  a(m(x,y),t) = a(x,a(y,t))
+module Examples
+export module_pres
+
+using ..Theories
 
 module_pres = Presentation(
   2,
@@ -10,3 +10,5 @@ module_pres = Presentation(
     lhs=Appl(2, [Appl(1, [Var(1), Var(2)]), Var(3)]),
     rhs=Appl(2, [Var(1), Appl(2, [Var(2), Var(3)])]))]
 )
+
+end
