+++ test/runtests.jl
@@ -1,6 +1,7 @@
 using AlgThys
+using AlgThys.Examples
 using Test
 
-@testset "AlgThys.jl" begin
-    # Write your tests here.
+@testset "AlgThys" begin
+  @test check_presentation(module_pres)
 end
